Feature/#23

+++ HEW2023/Form6.cs
@@ -20,10 +20,17 @@ namespace HEW2023
         Dummy dummy = new Dummy();
         DataTable dt = new DataTable();
 
+        List<int> generateList = new List<int>();
         List<int> dataIndexList = new List<int>();
 
         private void Form6_Load(object sender, EventArgs e)
         {
+            DataGridView.AutoGenerateColumns = true;
+            //generateList = new List<int>();
+            dataIndexList = new List<int>();
+            dt = new DataTable();
+            dummy = new Dummy();
+
             this.FormBorderStyle = FormBorderStyle.FixedSingle;
             this.MaximizeBox = false;
 
@@ -38,10 +45,11 @@ namespace HEW2023
             DataGridView.ReadOnly = true;
             this.DataGridView.SelectionMode = DataGridViewSelectionMode.FullRowSelect;
 
-            //DataGridViewColumn column = DataGridView.Columns[0];
-            //column.Width = 120;
 
             DataGridView.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.None;
+            this.FormBorderStyle = FormBorderStyle.FixedSingle;
+            this.MaximizeBox = false;
+            DataGridView.AllowUserToAddRows = false;
 
 
             //DataGridView1の列の幅をユーザーが変更できないようにする
@@ -68,6 +76,8 @@ namespace HEW2023
             {
                 dt.Columns.Add(columnsList[i]);
             }
+
+            int generateCount = 0;
 
             //RowIndex
             for (int j = 0; j < dataCount; j++)
@@ -105,12 +115,14 @@ namespace HEW2023
                             dr[columnsList[k].ToString()] = dataList[j][k];
                         }
                     }
+                    //ブックマークtemp処理
+                    if (originalDataList[j][9] == "1" && originalDataList[j][8] == "")
+                    {
+                        generateList.Add(generateCount);
+                    }
                     dt.Rows.Add(dr);
-                }
-                else
-                {
-                    continue;
-                    dummy.StringDebug("delete_flag_1_" + (j+1));
+
+                    generateCount++;
                 }
             }
             DataGridView.DataSource = dt;
@@ -123,6 +135,19 @@ namespace HEW2023
             DataGridView.Columns[0].Width = 265;
             DataGridView.Columns[3].Width = 70;
             DataGridView.Columns[5].Width = 65;
+
+            //ソート無効化
+            foreach (DataGridViewColumn c in DataGridView.Columns)
+            {
+                c.SortMode = DataGridViewColumnSortMode.NotSortable;
+            }
+
+            //ブックマーク処理
+            foreach (int i in generateList)
+            {
+                dummy.intDebug(i);
+                DataGridView.Rows[i].DefaultCellStyle.BackColor = Color.Aquamarine;
+            }
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -142,24 +167,79 @@ namespace HEW2023
             String title = "削除しますか？";
             String message = "選択された「" + selectTitle + "」を削除しますか？";
 
-            if (dummy.selectMessageBox(dummy.MessageBox_re(title, message)))
+            bool result = false;
+
+            dummy.StringDebug("mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm");
+            Console.WriteLine(generateList.Contains(selectedRowIndex));
+            dummy.StringDebug("mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm");
+
+            if (generateList.Contains(selectedRowIndex))
             {
-                selectedRowIndex++;
-                if (dummy.sqlExectionQuery(deleteQuery(selectId)))
+                if (dummy.selectMessageBox(dummy.MessageBox_re("ブックマーク確認", "ブックマーク済みですが削除処理に進みますか？")))
                 {
-                    title = "削除完了";
-                    message = "選択された「" + selectTitle + "」を削除しました。";
-                    dummy.MessageBox_(title, message);
-                    this.Close();
-                }
-                else
-                {
-                    dummy.StringDebug("form6のquery実行にてエラー発生。");
-                    this.Close();
-                    return;
+                    if (dummy.selectMessageBox(dummy.MessageBox_re(title, message)))
+                    {
+                        if (dummy.sqlExectionQuery(deleteQuery(selectId)))
+                        {
+                            title = "削除完了";
+                            message = "選択された「" + selectTitle + "」を削除しました。";
+                            dummy.MessageBox_(title, message);
+                            generateList.Clear();
+                            Form6_Load(null, EventArgs.Empty);
+                        }
+                        else
+                        {
+                            dummy.StringDebug("form6のquery実行にてエラー発生。");
+                            this.Close();
+                            return;
+                        }
+                    }
                 }
             }
+            else
+            {
+                if (dummy.selectMessageBox(dummy.MessageBox_re(title, message)))
+                {
+                    if (dummy.sqlExectionQuery(deleteQuery(selectId)))
+                    {
+                        title = "削除完了";
+                        message = "選択された「" + selectTitle + "」を削除しました。";
+                        dummy.MessageBox_(title, message);
+                        generateList.Clear();
+                        Form6_Load(null, EventArgs.Empty);
+                    }
+                    else
+                    {
+                        dummy.StringDebug("form6のquery実行にてエラー発生。");
+                        this.Close();
+                        return;
+                    }
+                }
+            }
+                
         }
+
+        //private void hogehoge(String title, String message)
+        //{
+        //    if (dummy.selectMessageBox(dummy.MessageBox_re(title, message)))
+        //    {
+        //        if (dummy.sqlExectionQuery(deleteQuery(selectId)))
+        //        {
+        //            title = "削除完了";
+        //            message = "選択された「" + selectTitle + "」を削除しました。";
+        //            dummy.MessageBox_(title, message);
+        //            generateList.Clear();
+        //            Form6_Load(null, EventArgs.Empty);
+        //        }
+        //        else
+        //        {
+        //            dummy.StringDebug("form6のquery実行にてエラー発生。");
+        //            this.Close();
+        //            return;
+        //        }
+        //    }
+        //}
+
         private String deleteQuery(int selectBookId)
         {
             String deleteQuery = "UPDATE `books_list` SET `delete_flag` = '1' WHERE `books_list`.`id` = " + selectBookId.ToString();
